+++ javaproperties/reading.py
@@ -104,19 +104,6 @@ def parse(fp):
         else:
             yield (unescape(line), '', source)
 
-_unescapes = {'t': '\t', 'n': '\n', 'f': '\f', 'r': '\r'}
-
-def _unesc(m):
-    esc = m.group(1)
-    if len(esc) == 1:
-        return _unescapes.get(esc, esc)
-    else:
-        return unichr(int(esc[1:], 16))
-
-def _unsurrogate(m):
-    c,d = map(ord, m.group())
-    return unichr(((c - 0xD800) << 10) + (d - 0xDC00) + 0x10000)
-
 def unescape(field):
     """
     Decode escape sequences in a ``.properties`` key or value.  The following
@@ -136,4 +123,29 @@ def unescape(field):
     :rtype: text string
     """
     return re.sub(r'[\uD800-\uDBFF][\uDC00-\uDFFF]', _unsurrogate,
-                  re.sub(r'\\(u[0-9A-Fa-f]{4}|.)', _unesc, field))
+                  re.sub(r'\\(u.{0,4}|.)', _unesc, field))
+
+_unescapes = {'t': '\t', 'n': '\n', 'f': '\f', 'r': '\r'}
+
+def _unesc(m):
+    esc = m.group(1)
+    if esc[0] == 'u':
+        if not re.match(r'^u[0-9A-Fa-f]{4}\Z', esc):
+            # We can't rely on `int` failing, because it succeeds when `esc`
+            # has trailing whitespace or a leading minus.
+            raise InvalidUEscapeError('\\' + esc)
+        return unichr(int(esc[1:], 16))
+    else:
+        return _unescapes.get(esc, esc)
+
+def _unsurrogate(m):
+    c,d = map(ord, m.group())
+    return unichr(((c - 0xD800) << 10) + (d - 0xDC00) + 0x10000)
+
+class InvalidUEscapeError(ValueError):
+    def __init__(self, escape):
+        self.escape = escape
+        super(InvalidUEscapeError, self).__init__(escape)
+
+    def __str__(self):
+        return 'Invalid \\u escape sequence: ' + self.escape